Refactor: Organize core logic and examples into separate directories
