+++ package.json
@@ -6,16 +6,23 @@
   "scripts": {
     "build": "tsc",
     "start": "node dist/index.js",
-    "dev": "nodemon --exec ts-node src/index.ts",
+    "dev": "nodemon --exec ts-node src/examplesRunner.ts",
+    "examples": "ts-node src/examplesRunner.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": ["entity", "effects", "stats", "game", "api", "typescript", "interactions", "gear"],
+  "keywords": [
+    "entity",
+    "effects",
+    "stats",
+    "game",
+    "api",
+    "typescript",
+    "interactions",
+    "gear"
+  ],
   "author": "",
   "license": "ISC",
   "type": "commonjs",
-  "dependencies": {
-    "dotenv": "^17.2.2"
-  },
   "devDependencies": {
     "@types/node": "^24.5.2",
     "nodemon": "^3.1.10",
