+++ src/core/ActiveEffects.ts
