+++ src/core/EffectApplicator.ts
