+++ src/core/Entity.ts
