+++ src/core/FrameManager.ts
