+++ src/core/InteractionSystem.ts
