+++ src/core/OptimizedFrameSystem.ts
