+++ src/core/RNG.ts
