+++ src/core/RNGEffects.ts
